Feature/update create (#20)
Creating random and collect mixins, factoring the logic out to those. 
Updating the create functionality to reflect current practices.

---------
