+++ opensampl/mixins/__init__.py
@@ -1,0 +1,1 @@
+"""mixins to add additional functionality to base probes"""
