+++ opensampl/vendors/adva.py
@@ -13,12 +13,13 @@ from loguru import logger
 from pydantic import Field
 
 from opensampl.metrics import METRICS
+from opensampl.mixins.random_data import RandomDataMixin
 from opensampl.references import REF_TYPES
 from opensampl.vendors.base_probe import BaseProbe
 from opensampl.vendors.constants import VENDORS, ProbeKey
 
 
-class AdvaProbe(BaseProbe):
+class AdvaProbe(BaseProbe, RandomDataMixin):
     """ADVA Probe Object"""
 
     timestamp: datetime
@@ -32,7 +33,7 @@ class AdvaProbe(BaseProbe):
         r"(?P<hour>\d+)-(?P<minute>\d+)-(?P<second>\d+)\.txt(?:\.gz)?"
     )
 
-    class RandomDataConfig(BaseProbe.RandomDataConfig):
+    class RandomDataConfig(RandomDataMixin.RandomDataConfig):
         """Model for storing random data generation configurations as provided by CLI or YAML"""
 
         # Time series parameters
